ipa-replica-manage.new add new code to svn for ipa-replica-manage cli automation refactoring
git-svn-id: https://svn.devel.redhat.com/repos/ipa-tests/trunk@5429 aeb79ed8-e03d-0410-be94-af332b95afa2

+++ ipa-tests/beaker/ipa-server/acceptance/quickinstall/ipa-install.sh
@@ -719,9 +719,9 @@ ipa_install_prep_setTime()
 
 fixHostFile()
 {
-    if [ "$USEDNS" != "no" ]; then
-        ipa_install_prep_initVars
+    ipa_install_prep_initVars
 
+    if [ "$USEDNS" != "no" ]; then
         cp -af /etc/hosts /etc/hosts.ipabackup
         rlRun "sed -i s/$hostname//g    /etc/hosts"
         rlRun "sed -i s/$hostname_s//g  /etc/hosts"
@@ -738,9 +738,9 @@ fixHostFile()
 
 fixHostFileIPv6()
 {
-    if [ "$USEDNS" != "no" ]; then
-        ipa_install_prep_initVars
+    ipa_install_prep_initVars
 
+    if [ "$USEDNS" != "no" ]; then
         cp -af /etc/hosts /etc/hosts.ipabackup
         rlRun "sed -i s/$hostname//g    /etc/hosts"
         rlRun "sed -i s/$hostname_s//g  /etc/hosts"
@@ -755,9 +755,9 @@ fixHostFileIPv6()
 
 fixhostname()
 {
-    if [ "$USEDNS" != "no" ]; then
-        ipa_install_prep_initVars
+    ipa_install_prep_initVars
         
+    if [ "$USEDNS" != "no" ]; then
         if [ -f /etc/sysconfig/network ]; then
             if [ ! -f /etc/sysconfig/network-ipabackup ]; then
                 rlRun "cp /etc/sysconfig/network /etc/sysconfig/network-ipabackup"
@@ -795,19 +795,23 @@ rmIPv4addr()
 fixResolv()
 {
     ipa_install_prep_initVars
-    
-    # we use the RRTYPE here in $rrtype to determine if IPv4 vs IPv6 address needed.
-    if [ ! -f /etc/resolv.conf.ipabackup ]; then
-        rlRun "cp /etc/resolv.conf /etc/resolv.conf.ipabackup"
-    fi
-    if [ $(echo $MYROLE | egrep "REPLICA|CLIENT"|wc -l) -gt 0 ]; then
-        for ns in $(eval echo \$BEAKERMASTER_env${MYENV}) $(eval echo \$BEAKERREPLICA_env${MYENV}); do
-            nsaddr=$(dig +short $ns $rrtype|tail -1)
-            rlRun "echo \"nameserver $nsaddr\" >> /etc/resolv.conf.new"
-        done
-        rlRun "sed -i s/^nameserver/#nameserver/g /etc/resolv.conf"
-        rlRun "cat /etc/resolv.conf.new >> /etc/resolv.conf"
-        rlRun "rm -f /etc/resolv.conf.new"
+
+    if [ "$USEDNS" != "no" ]; then
+        # we use the RRTYPE here in $rrtype to determine if IPv4 vs IPv6 address needed.
+        if [ ! -f /etc/resolv.conf.ipabackup ]; then
+            rlRun "cp /etc/resolv.conf /etc/resolv.conf.ipabackup"
+        fi
+        if [ $(echo $MYROLE | egrep "REPLICA|CLIENT"|wc -l) -gt 0 ]; then
+            for ns in $(eval echo \$BEAKERMASTER_env${MYENV}) $(eval echo \$BEAKERREPLICA_env${MYENV}); do
+                nsaddr=$(dig +short $ns $rrtype|tail -1)
+                rlRun "echo \"nameserver $nsaddr\" >> /etc/resolv.conf.new"
+            done
+            rlRun "sed -i s/^nameserver/#nameserver/g /etc/resolv.conf"
+            rlRun "cat /etc/resolv.conf.new >> /etc/resolv.conf"
+            rlRun "rm -f /etc/resolv.conf.new"
+        fi
+    else
+        rlLog "USEDNS=no, skipping resolv.conf reconfig"
     fi
     rlRun "cat /etc/resolv.conf"
 }
@@ -815,19 +819,23 @@ fixResolv()
 fixResolvIPv6()
 {
     ipa_install_prep_initVars
-    
-    # we use the RRTYPE here in $rrtype to determine if IPv4 vs IPv6 address needed.
-    if [ ! -f /etc/resolv.conf.ipabackup ]; then
-        rlRun "cp /etc/resolv.conf /etc/resolv.conf.ipabackup"
-    fi
-    if [ $(echo $MYROLE | egrep "REPLICA|CLIENT"|wc -l) -gt 0 ]; then
-        for ns in $(eval echo \$BEAKERMASTER_env${MYENV}) $(eval echo \$BEAKERREPLICA_env${MYENV}); do
-            nsaddr=$(dig +short $ns $rrtype|tail -1)
-            rlRun "echo \"nameserver $nsaddr\" >> /etc/resolv.conf.new"
-        done
-        rlRun "sed -i s/^nameserver/#nameserver/g /etc/resolv.conf"
-        rlRun "cat /etc/resolv.conf.new >> /etc/resolv.conf"
-        rlRun "rm -f /etc/resolv.conf.new"
+        
+    if [ "$USEDNS" != "no" ]; then
+        # we use the RRTYPE here in $rrtype to determine if IPv4 vs IPv6 address needed.
+        if [ ! -f /etc/resolv.conf.ipabackup ]; then
+            rlRun "cp /etc/resolv.conf /etc/resolv.conf.ipabackup"
+        fi
+        if [ $(echo $MYROLE | egrep "REPLICA|CLIENT"|wc -l) -gt 0 ]; then
+            for ns in $(eval echo \$BEAKERMASTER_env${MYENV}) $(eval echo \$BEAKERREPLICA_env${MYENV}); do
+                nsaddr=$(dig +short $ns $rrtype|tail -1)
+                rlRun "echo \"nameserver $nsaddr\" >> /etc/resolv.conf.new"
+            done
+            rlRun "sed -i s/^nameserver/#nameserver/g /etc/resolv.conf"
+            rlRun "cat /etc/resolv.conf.new >> /etc/resolv.conf"
+            rlRun "rm -f /etc/resolv.conf.new"
+        fi
+    else
+        rlLog "USEDNS=no, skipping resolv.conf reconfig"
     fi
 }
